TierIV-Capstone-Backend/requirements.txt
# affine==2.4.0
# amqp==5.3.1
# annotated-types==0.7.0
# anyio==4.9.0
# appnope==0.1.4
# argon2-cffi==25.1.0
# argon2-cffi-bindings==21.2.0
# arrow==1.3.0
# asgiref==3.9.1
# asttokens==3.0.0
# async-lru==2.0.5
# async-timeout==5.0.1
# attrs==25.3.0
# babel==2.17.0
# beautifulsoup4==4.13.4
# billiard==4.2.1
# bleach==6.2.0
# blinker==1.9.0
# branca==0.8.1
# Brotli==1.1.0
# cached-property==1.5.2
# cachelib==0.13.0
# celery==5.5.3
# certifi==2025.4.26
# cffi==1.17.1
# charset-normalizer==3.4.2
# click==8.2.1
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# cligj==0.7.2
# colorama==0.4.6
# comm==0.2.2
# contourpy==1.3.2
# cycler==0.12.1
# debugpy==1.8.14
# decorator==5.2.1
# defusedxml==0.7.1
# dnspython==2.7.0
# et_xmlfile==2.0.0
# exceptiongroup==1.3.0
# executing==2.2.0
# fastjsonschema==2.21.1
# Flask==3.1.1
# Flask-Caching==2.3.1
# Flask-PyMongo==3.0.1
# Flask-SQLAlchemy-Lite==0.1.0
# folium==0.19.7
# fonttools==4.58.2
# fqdn==1.5.1
# geographiclib==2.0
# geopandas==1.1.0
# geopy==2.4.1
# greenlet==3.2.3
# h11==0.16.0
# h2==4.2.0
# hpack==4.1.0
# httpcore==1.0.9
# httpx==0.28.1
# hyperframe==6.1.0
# idna==3.10
# importlib_metadata==8.7.0
# importlib_resources==6.5.2
# ipykernel==6.29.5
# ipython==8.37.0
# isoduration==20.11.0
# itsdangerous==2.2.0
# jedi==0.19.2
# Jinja2==3.1.6
# joblib==1.5.1
# json5==0.12.0
# jsonpointer==3.0.0
# jsonschema==4.24.0
# jsonschema-specifications==2025.4.1
# jupyter-events==0.12.0
# jupyter-lsp==2.2.5
# jupyter_client==8.6.3
# jupyter_core==5.8.1
# jupyter_server==2.16.0
# jupyter_server_terminals==0.5.3
# jupyterlab==4.4.3
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.27.3
# kiwisolver==1.4.7
# kombu==5.5.4
# lxml==5.4.0
# mapclassify==2.8.1
# MarkupSafe==3.0.2
# matplotlib==3.10.3
# matplotlib-inline==0.1.7
# mistune==3.1.3
# munkres==1.1.4
# nbclient==0.10.2
# nbconvert==7.16.6
# nbformat==5.10.4
# nest-asyncio==1.6.0
# networkx==3.4.2
# notebook_shim==0.2.4
# numpy==2.2.6
# openpyxl==3.1.5
# osmnx==2.0.3
# overrides==7.7.0
# packaging==25.0
# pandas==2.3.0
# pandocfilters==1.5.0
# parso==0.8.4
# pexpect==4.9.0
# pickleshare==0.7.5
# pillow==11.2.1
# pkgutil_resolve_name==1.3.10
# platformdirs==4.3.8
# prometheus_client==0.22.1
# prompt_toolkit==3.0.51
# psutil==7.0.0
# ptyprocess==0.7.0
# pure_eval==0.2.3
# pycparser==2.22
# pydantic==2.11.7
# pydantic_core==2.33.2
# Pygments==2.19.1
# pymongo==4.10.1
# pyogrio==0.11.0
# pyparsing==3.2.3
# pyproj==3.7.1
# PySide6==6.9.1
# PySide6_Addons==6.9.1
# PySide6_Essentials==6.9.1
# PySocks==1.7.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.1
# python-json-logger==2.0.7
# pytz==2025.2
# PyYAML==6.0.2
# pyzmq==26.4.0
# rasterio==1.4.3
# redis==6.2.0
# referencing==0.36.2
# requests==2.32.3
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rio-vrt==0.3.1
# rpds-py==0.25.1
# scikit-learn==1.7.0
# scipy==1.15.2
# Send2Trash==1.8.3
# setuptools==80.9.0
# shapely==2.1.1
# shiboken6==6.9.1
# six==1.17.0
# sniffio==1.3.1
# snuggs==1.4.7
# soupsieve==2.7
# SQLAlchemy==2.0.41
# stack-data==0.6.3
# terminado==0.18.1
# threadpoolctl==3.6.0
# tinycss2==1.4.0
# tomli==2.2.1
# tornado==6.5.1
# traitlets==5.14.3
# types-python-dateutil==2.9.0.20250516
# typing-inspection==0.4.1
# typing-utils==0.1.0
# typing_extensions==4.14.0
# tzdata==2025.2
# unicodedata2==16.0.0
# uri-template==1.3.0
# urllib3==2.4.0
# vine==5.1.0
# wcwidth==0.2.13
# webcolors==24.11.1
# webencodings==0.5.1
# websocket-client==1.8.0
# Werkzeug==3.1.3
# wheel==0.45.1
# xyzservices==2025.4.0
# zipp==3.23.0
# zstandard==0.23.0

# Filtered list below

celery==5.5.3
certifi==2025.4.26
Flask==3.1.1
Flask_Caching==2.3.1
flask_cors==6.0.1
flask_pymongo==3.0.1
flask_sqlalchemy_lite==0.1.0
geopandas==1.1.0
networkx==3.4.2
numpy==2.3.2
openpyxl==3.1.5
osmnx==2.0.5
pandas==2.3.1
pydantic==2.11.7
pymongo==4.13.2
python-dotenv==1.1.1
scipy==1.16.1
Shapely==2.1.1
tqdm==4.67.1
